Refactoring and minor fixes (#14)
* Fix ESLint config file for compatibility with newer version

Used by ESLint plugin in VSCode

* Update package-lock.json

* Fix syntax issue reported by ESLint

* Do not use named capture groups for compatibility with older JS version

* Improve naming of commands

* Improve code comments formatting

* Fix ESLint errors

* Reformat vue files

* Reformat CSS file

* Rename CSS class for player control buttons

* Remove unused functionality of upstream repo plugin

e.g. command and UI to add/delete comment, small player UI

* Remove more unused functionality

* Fix spacing error in vue template that broke build

* Refactor and clarify logic of calculating comment 'rank'

* Clear event listeners before unmounting

* Cosmetic changes to comment formatting

* Fix missing initialization of looping state from audio HTMLElement

* Revert "Clear event listeners before unmounting"

This reverts commit ec1877b9b03daaa5891ad3793efc046de0dcdc9c.

Event listeners are not to be cleared after the player
is removed from the DOM, because we still need to listen
for events triggered by commands.

* Fix skipback and loop button not setting audio to current

As a result of the bug, when clicking on the skipback or
loop button on a player different than the current one,
the currently playing track would be skipped back or looped
instead.

* Fix skip back button not working if audio is not current

* Do not toggle play when seeking to timestamp

* Minor cleanup and comment improvements